Refactored the socket handlers
